requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# src/data/make_dataset.py: 2
Click == 7.0

# src/func/geo_utils.py: 1
Fiona == 1.7.13

# src/func/geo_utils.py: 2
Shapely == 1.6.4.post2

# src/func/gather_tweets.py: 12
# src/func/geo_utils.py: 3
geopandas == 0.4.0

# src/func/gather_tweets.py: 25
# src/func/labmtgen.py: 6
numpy == 1.16.4

# src/func/gather_tweets.py: 18
# src/func/labmtgen.py: 20
# src/func/regex.py: 12
# src/func/tweet_utils.py: 8
pandas == 0.25.0

# src/func/gather_tweets.py: 11,31
# src/func/mongo_utils.py: 1
pymongo == 3.7.2

# src/data/make_dataset.py: 5
python_dotenv == 0.12.0

# src/func/regex.py: 13
requests == 2.22.0

# setup.py: 1
setuptools == 45.2.0

# src/func/labmtgen.py: 2
# src/func/tweet_utils.py: 2
ujson == 1.35
